// projects/zomatoRestoListApp/src/interface.ts
export interface R {
    has_menu_status: HasMenuStatus;
    res_id: number;
}

export interface Location {
    address: string;
    locality: string;
    city: string;
    city_id: number;
    latitude: string;
    longitude: string;
    zipcode: string;
    country_id: number;
    locality_verbose: string;
}

export interface Title {
    text: string;
}

export interface BgColor {
    type: string;
    tint: string;
}

export interface RatingObj {
    title: Title;
    bg_color: BgColor;
}

export interface UserRating {
    aggregate_rating: string;
    rating_text: string;
    rating_color: string;
    rating_obj: RatingObj;
    votes: string;
}
// ...
export interface Photo {
    photo: Photo2;
}

export interface Review {
    review: any[];
}

export interface AllReviews {
    reviews: Review[];
}
// ...
export interface ZomatoEvent {
    event: Event;
}

export interface Restaurant2 {
    R: R;
    apikey: string;
    id: string;
    name: string;
    url: string;
    location: Location;
    switch_to_order_menu: number;
    cuisines: string;
    timings: string;
    average_cost_for_two: number;
    price_range: number;
    currency: string;
    highlights: string[];
    offers: any[];
    opentable_support: number;
    is_zomato_book_res: number;
    mezzo_provider: string;
    is_book_form_web_view: number;
    book_form_web_view_url: string;
    book_again_url: string;
    thumb: string;
    user_rating: UserRating;
    all_reviews_count: number;
    photos_url: string;
    photo_count: number;
    photos: Photo[];
    menu_url: string;
    featured_image: string;
    medio_provider: any;
    has_online_delivery: number;
    is_delivering_now: number;
    store_type: string;
    include_bogo_offers: boolean;
    deeplink: string;
    is_table_reservation_supported: number;
    has_table_booking: number;
    book_url: string;
    events_url: string;
    phone_numbers: string;
    all_reviews: AllReviews;
    establishment: string[];
    establishment_types: any[];
    order_url: string;
    order_deeplink: string;
    zomato_events: ZomatoEvent[];
}

export interface Restaurant {
    restaurant: Restaurant2;
}

export interface RestaurantList {
    results_found: number;
    results_start: number;
    results_shown: number;
    restaurants: Restaurant[];
}

export interface RequestPayloadForRestoList{
    listCount: string | number;
    latitude: string;
    longitude: string;
    collectionId: string | number ;
}